requirements.txt
Django
django-crispy-forms
django-simple-select2>0.1.0
